Controlling the last activity not to be deleted

+++ src/services/activities/controllers/delete.js
@@ -4,8 +4,14 @@ const { ActivitiesModel } = require('../../../database/models');
 const controller = async (req, res)=>{
   console.log('/activities - DELETE');
   try{
-    const activity = await ActivitiesModel.deleteOne({ _id: req.params.id });
-    if (!activity.deletedCount) {
+    const activity = await ActivitiesModel.findById(req.params.id)
+    const proyectActivities = await ActivitiesModel.find({projectId: activity.projectId});
+    if (proyectActivities.length === 1){
+      console.log('/activities - Method Not Allowed');
+      return res.status(405).send();      
+    }
+    const deleted = await ActivitiesModel.deleteOne({ _id: req.params.id });
+    if (!deleted.deletedCount) {
       console.log('/activities - Not Found');
       return res.status(404).send();
     }
